bug fix for transport code
Now plant uptake of volatile tracers (exclude NH3 and NH3B) are applied
as fast transport, while non-volatiles are applied as slow transport.
Consequently, the CO2 production and surface efflux difference is reasoanble.
However, because root-soil gas exchange and in soil tracer transport are done
separately, difference are accumulated as transport error for diagnosing
the quality of tracer transport.