Running ATSEcoSIM Driver with Prescirbed Phenology
This commit contains several fixes to the initialization of the roots
in the ATSEcoSIM driver. The fixes can be broadly broken up into
two groups. Setting soil variables needed in Uptakesmod, these
are set in ATSEcoSIMAdvanceMod. And variables related to the root
properties themselves, set directly in UptakesMod.F90. The soil
properties are imported from ATS, while the root properties are set
manually for now, based off the standalone satellite phenology test.