Working Albedo modification in phenology
Fixed a bug where the sw radiation was not computed correctly. So now
The Albedo calculation is actually doing something. modifying the
albedo by hand does produce the expected result, lower albedo leads
to faster melting and less snowpack, but futher testing is needed.
This has not been tested with the phenology yet.