Working canopy, root, and snowmelt
This update fixes a few remainig issues with the canopy initialization.
The canopy requires canopy variables LWRadCanGPrev_col, TLEX_col, and
TSHX_col to be carried over between timesteps. If this is not done,
the snow fails to melt. Now the snow melts, and responds in a
reasonable way to shading. As an additional bonus the snow seems
to go isothermal while melting as epected, although more testing
is underway. Additionally root uptake appears to be occuring, but
again, further testing is needed.